Refact : storybook globalTheme 적용 안되는 에러 해결
