+++ finda2.0/src/components/common/Poster/type.ts
@@ -1,0 +1,4 @@
+export type posterType = {
+  title: string;
+  src: string;
+};
